# packages/chat-server/docs/api.swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "API Title",
    "version": "1.0"
  },
  "servers": [{ "url": "http://localhost:8080" }],
  "paths": {
    "/ping": {
      "description": "ping",
      "get": {
        "summary": "ping",
        "operationId": "ping",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/text": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/translate": {
      "description": "translate",
      "post": {
        "summary": "translate",
        "operationId": "translate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  }
                },
                "required": ["text", "from", "to"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/text": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "description": "chat",
        "operationId": "chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "enum": ["user", "assistant"]
                    },
                    "content": {
                      "type": "string"
                    }
                  },
                  "required": ["role", "content"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/chat-stream": {
      "post": {
        "description": "chat",
        "operationId": "chatStream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "enum": ["user", "assistant"]
                    },
                    "content": {
                      "type": "string"
                    }
                  },
                  "required": ["role", "content"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/get-region-and-token": {
      "description": "getRegionAndToken",
      "get": {
        "summary": "getRegionAndToken",
        "operationId": "getRegionAndToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "region": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": ["region", "token"]
                }
              }
            }
          }
        }
      }
    }
  }
}